// packages/orgs/src/scratch/types.ts
export interface ScratchOrg {
  accessToken?: string;
  alias?: string;
  elapsedTime?: number;
  expiryDate?: string;
  failureMessage?: string;
  instanceURL?: string;
  isScriptExecuted?: boolean;
  loginURL?: string;
  orgId?: string;
  password?: string;
  recordId?: string;
  sfdxAuthUrl?: string;
  signupEmail?: string;
  status?: string;
  tag?: string;
  username?: string;
}